spec: create /etc/leapp/actor_conf.d
Add additional build steps to the specfile that create the actor
configuration directory. The directory is owned by the package, so
it gets removed when the user uninstalls leapp.

Also prepared some comment lines for future when we will want to
include some configuration files as part of the rpm.