kernelcmdlineconfig: Use args from first entry when multiple entries are listed
Instead of erroring out when grubby lists multiple entries for the
default kernel, always use the `args=` and `root=` from the first one and create
a post-upgrade report. The report instruct user to ensure those are the
correct ones or to correct them.

This can happen, for example, if MAKEDEBUG=yes is set in
/etc/sysconfing/kernel.

Jira: RHEL-46911