Add el9toel10 actor to handle symlink -> directory with ruby IRB.
The `/usr/share/ruby/irb` path is a symlink in RHEL 9,
but a regular directory in RHEL 10.
This puts us back in line with RHEL 8 and Fedora in terms of the
path's file type regarding the rubygem-irb package.

Since this was not handled on RPM level, handle it as actor again.
This was copied and adjusted from same-named el8->el9 actor.

We do not care about the validity or target of the symlink, we just
remove it to allow DNF create the correct directory on upgrade.

Without this workaround, the upgrade will fail in transaction test with
reports of file conflicts on the directory path.

Users should not expect to ever retain anything in this directory.