Verify supported target OS version in actors
Originally when user specified the target system release using
`--target` CLI option the verification has been performed immediately
as only supported releases have been listed as possible choices for
this option. The benefit of this solution was that users did not have
to wait for all other checks to realize they execute leapp probably
incorrectly. Unfortunately,
 * number of users do not understand why only some versions are supported
 * users upgrading with via various webUIs presenting only leapp reports
   could not see the error message available in terminal

To resolve this problem the checks are moved into actors so in case
of specified unsupported target version the information is present
in generated leapp reports.

Current behaviour is like this:
 * in case of invalid input (incorrect format of input data) the hard
   error is raised as before immediately. Malformed input data will
   not be processed anyhow by any actors
 * report error when the specified target major version is not direct
   successor of the current system version. I.e. specify 10.0 when
   upgrading from RHEL 8 (only RHEL 9 is acceptable).
   * this prevents number of cryptic errors as actors are not prepared
     for this situation
 * report standard inhibitor if the target release is not in the defined
   upgrade path, unless LEAPP_UNSUPPORTED=1
 * running leapp in unsupported (devel) mode skips the inhibitor and
   entire report

Additional changes:
* Update error message when format of target version is incorrect to
  clarify the expected version format

jira: RHEL-51072

Co-authored-by: Petr Stodulk <pstodulk@redhat.com>