Use --sysinit when calling vgchange from mount_usr.sh
The mount_usr.sh script runs 'lvm vgchange': if there are logical
volumes present that require monitoring (snapshots, thin pools, RAID,
etc.) the command will attempt to launch dmeventd.

Since dmeventd is not installed in the dracut initramfs this produces a
warning and causes the lvm command to exit with non-zero exit status
even though the volume group has been activated and LV block devices are
available.

This in turn triggers the retry logic in mount_usr.sh: once the retries
are exhausted the script carries on and successfully initiates the
upgrade process.

The --sysinit switch is used by the LVM dracut modules for this reason.
From vgchange(8):

       --sysinit
        Indicates  that  vgchange/lvchange  is being invoked from early
        system initialisation scripts (e.g. rc.sysinit or an initrd),
        before writable filesystems are available. As such, some
        functionality needs to be disabled and this option acts as a
        shortcut which selects an appropriate set of options. Currently,
        this is equivalent to using  --ignorelockingfailure,
        --ignoremonitoring,  --poll  n, and setting env var
        LVM_SUPPRESS_LOCKING_FAILURE_MESSAGES.  vgchange/lvchange skip
        autoactivation, and defer to pvscan autoactivation.

Testing with this change I no longer see the long delay booting the
upgrade initramfs when snapshot LVs are present.