Introduce deprecated IPUPaths msg (temporary solution)
This is hackish precursor to move checking of the specified
target system versions into actors to be able to create report
when unsupported target version is specified.

The problem is that currently there is no information about the
supported upgrade paths in messages. Also, the information about
supported source and target versions are stored in two different
places (in system upgrade common repo):
 * shared configs.version library
 * files/upgrade_paths.json

As a temporary solution let's introduce IPUPaths message which
will contain filtered data from the json file based on:
 * the upgrade flavour (default, saphana)
 * and source major version

There is no value to print information to users about different
upgrade paths for other flavours and OS major versions. The model
is marked as deprecated so we can remove it in the next release when
we redesign this solution to unify how actors get this data
(and define them just in one place).

jira: RHEL-51072