pytest: update conftest to load leapp.snactor.fixture explicitly
The snactor plugin has been originally specified between entrypoints
in the leapp python package. However, this has been breaking pytest
on systems where snactor has not been installed (as pytest as very
interesting design...). So the pytest plugin hook had to be removed.
As it has no other value than for running unit-tests using pytest,
let's just specify it in the conftest.py file explicitly for running
the tests as it should actually be all the time.