livemode(cfg): use framework's configurability instead of ad-hoc INI
Drop the ad-hoc INI-based config for livemode in favour and replace it
with configuration facilitated by the leapp framework. The list of
configuration options remains (and their semantics) remains almost
unchanged, save for some field names that are renamed in a way that
better reveals their effect to the user.

Jira-ref: RHELMISC-10648