feat(livemode): remove the use of LEAPP_DEVEL_ENABLE_LIVEMODE
The environmental variable has been introduced to prevent accidental
execution of livemode. However, in order for users to use the feature,
this environmental variable introduced unnecessary friction. Therefore,
this patch removes the use of the variable. Instead, whitelisting
experimental actors that facilitate livemode should be the only
mechanism that is used to enabled/disable the feature.

Jira-ref: RHELMISC-10648