DOC: rename configuring-ipu.md to envars.md
The file has bas been created before we introduced the section of
"Configuring in-place upgrade" and it contained mixture of various
stuff. We have the section now, so let's rename this file to envars
and keep here documented environments variables only.

Dropping information from this file about actors' configurations.
This needs to be anyway documented separately now and original
information has not been so helpful.