repomap: Process repositories based on distro
The original solution expect just IPU of RHEL systems and so only
source RHEL DNF repositories have been mapped to the target RHEL
repositories. Making this more distro agnostic for CentOS-like
systems, the mapping needs to be updated to respect distro specific
repositories.

Hence a 'distro' field has been added in the 1.3.0 version of the
repomap.json format. The field is required for all the `repositories`
entries. RepomapDataHandler now takes in distro as a constructor
parameter and only consider entries with the matching distro value.

Reflects also changes in the repomap json schema for version 1.3.0,
which drops enumeration for the `rhui` field and replace it by simple
String type to make the solution more scalable. This also means that
vendor of the repomap data file must validate this field "manually".

All related unit-tests have been updated. As the commit is already
large, keeping the update of data files on a separate commit.

Update the PESIDRepositoryEntry model:
 * add the distro field represented by String
 * change rhui type from StringEnum to String

Jira: RHEL-80336

WIP setuptarget repos tests for distro

Add repositories mapping unit tests for distro