CI: Unit-tests: Rename tasks & tests executions
Previous name have been very long, repeating, and it has been
problem to read the whole informatino about what tests are performed
as the description has been shortened in WebUI, e.g.:

  Unit Tests / Run unit tests in containers (Run python linters for el8toel9 and common repositories on p...

Let's change the name so it's shortened and easier read what test
has been actually performed. Also, no need to mention that tests are
running in container anymore, as we are not running unit-tests outside
of the container in upstream. Example of new style names:

  Unit Tests / Run: Linters (python:3.9; repos:el9toel10,common) (pull_request)
  Unit Tests / Run: Unit Tests (python:3.9; repos:el9toel10,common) (pull_request)