Fix target version format checks
Currently only the format of a version specified by
LEAPP_DEVEL_TARGET_RELEASE is checked, when specified via --target
cmdline argument it isn't, which is a bug.

This patch fixes the bug, by enabling the check for --target too.

NOTE: these are only format checks, the state of support is already
checked later in the upgrade process by the checktargetversion actor.

Jira: RHEL-96238