Adjust DNF stream variable
CentOS Stream systems the `stream` DNF variable is used instead
of the `releasever`, which cause an issue when creating target repos,
becuase the $stream variable is handled incorrectly. This patch is
adjusting the content of /etc/dnf/vars/stream in scratch container,
so $stream var is replaced with correct target version of system.

Jira: RHEL-95982