Inhibit upgrade on systems with cgroupsv1 on 9to10 (#1392)
* Inhibit upgrade on systems with cgroupsv1 on 9to10

cgroups-v1 were deprecated in RHEL 9 and are unsupported in RHEL 10. The
option to switch to cgroups-v1 is ignored. Users have to migrate to
cgroups-v2 in order to upgrade.

Both legacy and hybrid modes are unsupported. Only unified hierarchy
(cgroups-v2) is supported. More info on how the modes are switched is
at:
https://www.freedesktop.org/software/systemd/man/247/systemd.html#systemd.unified_cgroup_hierarchy
and
https://www.freedesktop.org/software/systemd/man/247/systemd.html#systemd.legacy_systemd_cgroup_controller.

Jira: RHEL-81212