feat: scan source system's default boot entry
Add an actor that scans the default boot entry by calling grubby. New
model named 'DefaultSourceBootEntry' is introduced, allowing to share
obtained information with actors. At the moment, only the initramfs path
and kernel path are scanned, the rest of the default boot entry's
information is ignored.

(Done in preparation of:)
Jira-ref: RHEL-102591