userspacegen: Add repo gathering for non-RHEL distros
Extend the _get_rh_available_repoids() function, which is renamed to
_get_distro_available_repoids(), to work with other distros. The
idea is the same as for RHEL, scan well-known distro-provided repofiles
for repositories.

Also the TargetRepositories.rhel_repos is replaced by
TargetRepositories.distro_repos.

These changes together enable the use of repomapping on non-RHEL repos,
as before this change the --enablerepo option had to be used to specify
target repos and they were treated as custom repos.

TODO
- The well-known distro-provided repofiles will have to be hardcoded
somewhere
- The function should probably completely replace
  rhsm.get_available_repo_ids()

Also, the unused _get_rhui_available_repoids() function is removed.