Makefile: Make sanity-check respect $(REPOSITORIES)
Since b6e84f7, the sanity check, runs in each repository and doesn't
respect $(REPOSITORIES). This breaks the sanity-check when it's run with
a version of Python which is incompatible with the code in particular
repository.

For example with python3.6 and el9toel10 repo, python crashes with the
following because python3.6 doesn't understand the type hint:
  File "/payload/repos/system_upgrade/el9toel10/actors/mysql/scanmysql/libraries/scanmysql.py", line 35, in <module>
    def _check_incompatible_config() -> set[str]:
TypeError: 'type' object is not subscriptable