Rename Containerfiles to consistent names
The container files (both those used in CI and locally by the Makefile)
and both build and test, use various base containers - Centos, UBIs.

This patch unifies the naming scheme to Containerfile.elX to avoid
making the user remember which one is used by which command an at
which version.

The containerfiles used by Github Actions are moved to separate
folder to avoid name clashes.

Also, in the GitHub actions test workflow the
--security-opt=seccomp=unconfined is removed. Not sure why it was set in
the first place, but seems it's not needed anymore.