LiveMode: Add /etc/crypttab file to the target userspace container
When upgrading with LiveMode, the auto-unlock of encrypted devices
fails because the /etc/crypttab configuration file is not present inside
the squashfs, causing the boot process to fail.

This change copy /etc/crypttab file to the target userspace container
during the upgrade process, allowing encrypted devices to be properly
unlocked.

Jira: RHEL-90098