lib/version: Also peform version correction when target is centos
Currently the version "autocorrection" is performed only when the source
system is centos, however we might get a major only version even when
the source doesn't use them and the target does. For example when
matching whether a version is between source and target version as is
done in peseventsscanner.