squash! feat(mount_unit_gen): generate mount units for the upgrade initramfs
@pstodulk:
Updated the code to cover also other systemd targets that can be
covered by systemd-fstab-generator. Also cover the situation when
a directory with systemd target (requires, wants) already exists.
Tests have been updated.

Note that there are still possible issues hidden in the generate
mount unit files as we update at this moment just the `Where` clause
however we are not touching anything else. (Before, After,
RequiresMountsFor, ...). But keeping that for future development and
testing.

Co-authored-by: Petr Stodulka <pstodulk@redhat.com>

that's fun...