cs->rhel: Add hint to check target version when failed to create target userspace
This can happen when the determined target version of RHEL has not been
release yet. This happens when we already need to put the not yet
released version into upgrade_paths.json and a build into Centos Stream.

Also, move the existing proxy hints from 'details' to 'hint', otherwise
the original exeception message, which is in 'details' is overwritten.

TODO:
- Allow setting the one versions lower than the determined one.