Makefile: Do not copy unnecessary files to test containers
In addition to tut/, docs/, packaging/, .git/ and all the __pycache__
directories (and .pyc files) are excluded. This makes for a significant
speedup in container setup (output with -v added to rsync) as a lot less
needs to be copied:

Without this patch:
sent 2,405,488 bytes  received 19,754 bytes  194,019.36 bytes/sec
total size is 165,333,162  speedup is 68.17

With this patch:
sent 551,179 bytes  received 4,067 bytes  100,953.82 bytes/sec
total size is 23,280,513  speedup is 41.93

Some other small files and directories are still unnecessarily copied,
but those don't really affect the copied size that much.