Makefile: Properly copy commands dir to test containers
During a tests run, the commands/ directory is not used when running
commands tests. Instead they are imported from
tut/lib/python3.X/site-packages/leapp/cli/commands/ where python3.X is
the python used in the tut/ virtualenv. This directory is not updated on
it's own, so the let's copy the commands/ dir there instead.