Replace distro specific packages during conversion
There are certain packages that are distribution specific and need to be
replaced in the DNF upgrade transaction with their target distro
counterpart when converting during the upgrade. For example the release
and logos packages. Some packages, such as packages containing
repository definitions or GPG keys, need to be removed without any
replacement.

This patch introduces a new convert/swapdistropackages actor to
accomplish this. Currently only packages that need to be handled during
CS->RHEL and AL->RHEL conversion are handled, however the actor contains
a config dict to easily add more paths.

Jira: RHEL-110568