handle multipath devices in upgrade initramfs
Previously, multipath configs were scanned (and optionally) modified
only for the 8>9 upgrade paths. Known backward-compatibility issues were
fixed in the ApplicationsPhase, but any support for mounting multipath
devices was lacking. This patch refactors the handling of multipath so
that:
1) all multipath-related configs are copied into target uspace
2) any changes done in the ApplicationsPhase are performed to these
   copies instead
3) our modified multipath configs are included into the initramfs
4) dracut 'multipath' module is enabled when building upgrade initramfs
5) system configs are patched so that in the end they will look the same
   as if modified in the ApplicationPhase. This is achieved by simply
   copying our modified configs and replacing the old system ones.

These changes avoid reaching into the config-parsing code. An ugly
consequence of this is that some models (MultipathConfFacts8to9) needed
to be placed into the common repository, although they are only used for
8>9 upgrades. To address this ugliness would require designing and
producing a message that would contain the entire contents of the
primary multipath config (if we want to avoid parsing it twice).

jira-ref: RHEL-124306 (dev task)