Add new actor: scannvme: To scan NVMe setup
Introduce new model/message NVMEInfo containing information about
detected NVMe devices and basic configuration.

From the configuration it's loaded information about hostid and
hostnqn if defined. This configuration is mandatory for NVMe-oF
setups.

Devices are scanned directly from /sys/class/nvme/ directory, which
should contain information about all NVMe devices. The gathered
facts are limited to information important for the correct handling
of the IPU with NVMe devices - covering NVMe-FC.

Currently scanned information about each device:
    * device name
    * NVMe transport type
    * sys_path (/sys/class/...) to the device definition

If any of the information is not possible to obtain just log an error
message and skip the device - the issue is theoretical and it's not
expected to happen in real. But leaving log messages so in case of
issues we leave visible clues.

Jira-ref: RHEL-46807