Handle multipath configs during TargetTransactionFactsChecksTag
The target_uspace_configs and multipath_upgrade_conf_patcher actors
were originally executed during the ChecksPhase. However, as they
need to access the system storage, we do not want to keep them
part of the ChecksPhase at this moment. Moving them to
TargetTransactionFactsChecksTag instead. I know this use is not aligned
with the phase description, but we use it already even for other things
as well, so let's go rather this way this time.