multipath: do not crash when there is no multipath.conf
Our newly introduced handling of multipath in the upgrade initramfs has
a bug when it tries to check whether multipath_info.config_dir exists.
However, when multipath config does not exists, a default message with
multipath_info.config_dir=None is produced, causing an unhandled
exception. This patch fixes the issue. Moreover, an additional issue
when updated configs are not *guaranteed* to be placed into the target
uspace was discovered.